[rust png] Fix `blendAllRows` to avoid `subspan`ing beyond the buffer.
When `blendAllRows` is called on a animation frame that is a sub-rect
(e.g. 3x2) of the bigger image (e.g. 5x5), then:

1. The `srcFrame` and `dstFrame` need to accommodate only 5+3=8 pixels
2. The `rowStride` skips over 5 pixels.

Illustration (`x`s are the sub-rect, `x` and `+` are what needs to be
covered by `srcFrame` and `dstFrame`):

  .....
  .....
  .....
  .xxx+
  +xxx.

When looping over all the rows of `srcFrame`, the loop advances to the
next row by executing `srcFrame = srcFrame.subspan(rowStride)`.  This
call to `subframe` will assert and crash in the scenario above, because
we would index beyond the buffer (into the non-existent 6th row in the
example above).

This CL fixes this problem, by using `std::min` on the argument of
`subspan` (just as we already do in another `subspan` callsite in
`SkPngRustCodec.cpp`.

Fixed: 445556737
Change-Id: I991cc11e741b74712ec95cb96a9f7e0cff9788e3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1057676
Commit-Queue: Łukasz Anforowicz <lukasza@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>