[graphite] Add limit to number of paths added to small path atlas per flush
This is a short term solution to avoid regressions with the small path
atlas hybrid strategy. Longer term, the plan is to configure the
RasterPathAtlas differently depending on whether or not MSAA is available,
and then fall back to MSAA automatically if there's no room in the
atlas.

The limit of 256 appears to avoid regressions in MotionMark for a
variety of different minPathSizeForMSAA options.

Bug: b/441466076
Change-Id: Ic26bff39a739bf2050a124dec3d80c7bcb49ff70
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1058036
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>