Fix GrProgramDesc creation wrt dstReads
The fundamental issue was that the old code assumed we would only
ever do a dstRead in the Xfer Processor. This assumption falls apart for SkSL blend modes were we set the final XP to kSrc (in skpaint_to_grpaint_impl) and incorporate the dstRead into a shader (c.f. the destcolor GM). This broke key creation bc the needed dstRead bits wouldn't be set by the XP key creator.

The new code follows the structure of GrGLSLProgramBuilder::emitAndInstallProcs - where the emission order is:
   dstTexture
   Fragment Processors
   Xfer Processor

Bug: b/445572384
Change-Id: I9ce579945146b0f66d45c7b2b6da536ba6e14b04
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1058116
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>