[rustpng] Relax assert for setting sigBits
The `output_color_type`, which is the switch expression in `ToColor()`,
is the encoded color type transformed to add alpha value if the `tRNS`
chunk requires it [1]. So it is valid from the rust decoder that our
sigBit chunk span size may not match the number of components of our
`output_color_type`. So I think it is valid to just relax the ASSERT
here.

[1]
https://github.com/image-rs/image-png/blob/da8c376cee8a97512b5e0056039d0b5fe6250007/src/decoder/mod.rs#L626

Bug: 446395430
Change-Id: I4171639597b1d50d29501fef119d2253a49fe527
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1059637
Reviewed-by: Łukasz Anforowicz <lukasza@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>