Refactor build_tint.py code into shared file
This also changes how the .a file is created. Rather than taking all
the built .a archives and unzipping them and re-zipping them, we
create a dag of .o files in all the targets and dependencies and then
use ar to join them together. This will scale better to handling Dawn.

This also works better on Mac. Windows will need its own support,
of which we see a little bit in BUILD.gn

Bug: b/434977307
Change-Id: I856492d6786381e262830ce4e43b036000503582
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1059773
Reviewed-by: Brett Somocurcio <brettos@google.com>