[graphite] Add a ThreadSafeResourceProvider
This CL just adds the class and plumbs it in but doesn't make use of it yet. Each backend gets its own derived version of the ThreadSafeResourceProvider although only the Vulkan backend currently needs the derived version.

When ResourceCaches are thread safe the ThreadSafeResourceProvider can go away and the ResourceProvider held by the Context can be used.

Bug: b/434712686
Change-Id: Icbde64486dfb18a0f30adbeee696187da0fbc26a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1062136
Reviewed-by: Thomas Smith <thomsmit@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>