[graphite] Add findOrCreateScratchBuffer to ResourceProvider
In a follow-up CL, the current ScratchBuffer class will be
generalized to a BufferSubAllocator that can choose between scratch
and non-shareable Buffers. The current fReusableScratchStorageBuffers
field in BufferManager will be replaced with the "unavailable set"
that the scratch system uses in the ResourceCache.

Bug: b/339479804
Bug: b/330744081
Bug; b/263130762
Change-Id: Ib279f7e5013b7ae27fc44c3d0b323238ea17c3a5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1064418
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>