[graphite] Make use of PipelineCreateCacheControl in Vulkan
This CL centralizes the CreateGraphicsPipelines calls in create_graphics_pipeline and then uses the PipelineCreateCacheControl feature (if present) to determine if any new pipelines were actually generated.

Note that the addition of the VulkanSharedContext::pipelineCompileWasRequired facility requires that a non-const SharedContext be passed down.

Bug: b/448366723
Change-Id: Idde5d61d3bbb36cd20e69ed22524735bb6740ffe
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1064417
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>