[graphite] Relax circular corner detection for analytic clip
Corners are often not quite circular due to transformations. Since this
is applied in device space, many float errors are imperceptible so
adds a hefty tolerance for comparing X vs Y radius.

It also updates the logic for complex/9-patch. The extra if was
unnecessary, and technically could have caused a SkRRect::isRect case
from not being detected (assuming the clip stack or canvas hadn't
simplified it already).

Lastly, it fixes an issue where a complex or 9-patch RRect becomes
effectively simple after we apply the radii tolerance. However, in
this case `cornerFlags` was still equal to kAll and was getting mis
identified as an unsupported case of 3 or diagonal-only corners. This
counts the number of square corners and it'll be 0 in the good case,
and either 1 or 2 in the problem cases.

Bug: b/449273747
Change-Id: Ia3c3042a96e0f2a567ce5804ce6baf42343b8ca1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1066442
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: Thomas Smith <thomsmit@google.com>