Pass in targetSurface to asView for same surface check.
When drawing an image, we need to know if we're drawing into a surface
that is backend by the same texture. If so, we need to perform a copy
since it is not legal to have the same texture. In Ganesh the copy on
write mechanisms are all controled by the ProxyChooser. This change
plumbs down the targetSurface proxy into ProxyChooser::chooseProxy
where we will check if we're drawing to the same surface or not.

Bug: b/387306744
Change-Id: I96e80e58a965ee8b94d13736fd3f3526d8b5778f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/972596
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>