[graphite] Refactor ScratchBuffer into BufferSubAllocator
ScratchBuffer is renamed and adjusted to BufferSubAllocator with its
main new functionality being:
 - Can suballocate mapped ranges now if it was created from a mapped buffer
 - Can reset its base alignment after suballocations to flexibly have more control over data packing when the caller knows when binding will happen
 - Stores the BufferState index so it can be used for any allocation category in DrawBufferManager

Under the hood of DrawBufferManager, BufferSubAllocator is used to
replace prepareBindBuffer(), prepareMappedBuffer() and consolidate the
different getXWriter() functions. With how the getWriter() and
getBinding() functions are implemented, this shouldn't lead to any real
behavioral differences for now. The only place BufferSubAllocator is
used outside of DrawBufferManager is previous ScratchBuffer sites.
 - The one exception is how scratch buffers are now reused, which switches to using the ResourceCache scratch system and we'll return an available scratch buffer with an exact key match.

In follow-up CLs, the getXWriter() functions will be replaced with
factories for getting specific types of BufferSubAllocator so that the
uniform tracker and DrawWriter can use the allocator directly.

Bug: b/263135784
Bug: b/263130762
Bug: b/339479804
Bug: b/330744081
Change-Id: I65974079173e6ba09254bda4785563b732ff05b1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1065176
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>