[graphite] Use BufferSubAllocator in UniformTracker
This removes getUniformWriter, getSsboWriter, and getAlignedSsboWriter.
These are replaced with getMappedUniformBuffer and
getMappedStorageBuffer. They return a tuple that represents the first
suballocation from a BufferSubAllocator as well as the
BufferSubAllocator so that additional allocations can be made directly
against it.

Updates DispatchGroup and unit tests to call the new
getMappedStorageBuffer. Updates DrawPass to store the returned
BufferSubAllocator and sub allocate against that first, only querying
the DrawBufferManager for a new buffer when the allocation fails.

For the uniform and storage buffer types tracked by BufferManager, it is
no longer suballocating BufferState::fAvailableBuffer. Now
fAvailableBuffer only represents something that was released and could
be returned by another call to getBuffer() if there was room. An
important behavioral change from this is that now each UniformTracker
holds its own allocator, which means that paint uniforms will write to
an entirely different GPU buffer than renderstep uniforms. Previously
they would interleave within the same GPU buffer based on reuse and
availabel space.

Since the long term plan is to switch to a single buffer binding, this
change is temporary.

In a follow-up CL, the vertex/index writer functions will be replaced
with similar signatures to getMappedXBuffer and DrawWriter will hold on
to the returned BufferSubAllocator.

Bug: b/263130762
Bug: b/448858571
Change-Id: If15d66f5646477ce06774168c2e0e980db24b65d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1066550
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Thomas Smith <thomsmit@google.com>