[graphite] Flush Devices that depend on a given Device being flushed
Device::notifyInUse() now marks the Device as having been drawn into
another DrawContext. When this is true, we know it has to flush
transitive dependencies; if it's never been drawn into another device,
then that doesn't matter and it can be skipped.

Since this is only an issue for non-scratch devices, we don't set the
flag for scratch devices. Scratch devices do not go through cycles of
reads + writes.

In order to flush transitive dependencies, we call the new
flushTrackedDevices that takes a TextureProxy, and pass in the Device's
render target proxy. This is currently the only use of this more
restricted flushTrackedDevices, but we could expand it to include
flushing just the devices that depend on an atlas page for instance.

Adds a new NotifyInUse test that correctly failed before this change and
now passes with the change.

Bug: b/450860941
Change-Id: I9a4d82760846ed105c1839ebde6a7296e6d956e0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1071446
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: Thomas Smith <thomsmit@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>