Allow GMs to specify surface props
There are a number of gms which attempt to test the effect of pixel
geometry settings. These gms generally create a raster backend with the
pixel geometry desired and then blit that into the final canvas. This is
less than ideal as the intent is to test the current config. Instead,
allow a gm to override the pixel geometry and have dm and viewer ask for
this override when requested.

Change-Id: Ieb396b0ee478688ed269d3b156618ae879884217
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1072999
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>