[graphite] Use consistent relative circular test for SkRRects
There are a number of places in Graphite that require SkRRects to have
circular corners. This is different than being simple, which is that an
SkRRect have the same X,Y radii for each corner. For AllCornersCircular,
each corner can have a different radius, but locally it is effectively a
circle.

These places are:
- Device::drawGeometry and AnalyticRRectRenderStep for drawing a stroked round rect because orthogonal outsets on a circular corner produce another rounded rectangle, while that is not the case for stroked elliptical corners.
- Device::drawDRRect to determine that a outer/inner round rect combination is equivalent to a stroked round rect (same circular requirement as above).
- Analytic clipping in ClipStack, uses circular corners for simplicity in the math and the amount of data needed to be uploaded.

There is a very related check occurring in AnalyticBlurMask for blurring
round rects, which currently requires the rectangle to be simple and
circular. This is not being updated here because I think there is a
larger overhaul that can be done to AnalyticBlurMask's logic to avoid
having to restrict itself at all.

Lastly, this adds a new "relative" AllCornersCircular function to
SkRRectPriv, which produces more stable classifications when an SkRRect
is transformed by a scaling matrix. While Ganesh could use this in a few
places, I chose not update its call sites to avoid running into
unexpected issues.

Bug: b/449273747
Change-Id: I790123ccb94c6e1fa3f105f1a01dc6431189c367
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1073336
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>