Prevent Viewer::fDisplay from always being nullptr
In "Add graphite-specific version of tools/DisplayParams" [0]
DisplayParamsBuilder::build was changed from returning a copy of the
current DisplayParams to moving the current DisplayParams to the caller.
However, Viewer's constructor depended on calling `build` more than
once, first to set up the `fWindow` (the current mutable state of the
window) and then again to set the `fDisplay` (the immutable physical
properties of the display). The semantics of the class were changed to
cause the builder to be in a moved-from state, which could cause
undefined behavior if any other methods were called after `build`.

Change `build` to `detach` to better document what it does. Add asserts
that the builder is not used in the moved-from state. Change Viewer to
set `fDisplay` initially, the clone into `fWindow`.

[0] https://skia.googlesource.com/skia.git/+/10670fef5b2226131c79bb1f5b568a598f136226

Change-Id: I0ab9a1a026ed9f1c0244dfd6ade2f9974a7e8fe0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1074376
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>