Mark some private constructors explicit
This is considered good practice by the style guide.
https://google.github.io/styleguide/cppguide.html#Implicit_Conversions

I don't expect this to break any clients, but I've made it a separate
commit from the ones touching the public API.

I asked Gemini CLI to help me with this one. I had it use a regex
to find candidates and then it narrowed down the pool and applied
the edits.

Change-Id: I3b519698508b186438a4a185c92472dccfa15313
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1075876
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>