Add explicit keyword to many public single argument constructors
This is considered good practice by the style guide.
https://google.github.io/styleguide/cppguide.html#Implicit_Conversions

I asked Gemini CLI to help me with this one. I had it use a regex
to find candidates and then it narrowed down the pool and applied
the edits.

This could break clients if they were relying on the auto-conversion.

Change-Id: I25f9b294d34c2c967504c59f0389fddc27237dea
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1075696
Reviewed-by: Michael Ludwig <michaelludwig@google.com>