Speculative fix for flaky dawn builds
We are building both Dawn and Tint at the same time, which
are both using the same jinja source files. Maybe the .pyc
files are getting into an inconsistent state and that's
causing a race condition?

https://stackoverflow.com/a/154617

Also add some logging to make sure we are using the non-system
python3 executable.

Bug: b:434977307
Change-Id: If0f2f4a2521eb00c38c54175711788f62ca4a489
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1076956
Reviewed-by: Thomas Smith <thomsmit@google.com>
Reviewed-by: Eric Boren <borenet@google.com>