Use SkCanvas::baseRecorder for SkImage::isValid
In "Deprecate non-recorder SkImage methods" [0] some of the tests for
stored image validation were changed to source from
SkCanvas::baseRecorder (like for `drawbitmaprect`) but others were
changed to use SkCanvas::recordingContext()->asRecorder(). However, the
GrRecordingContext is only available with a Gr backend. The gms
continued to draw in dm because they were only drawn once (the fImage
was nullptr and was created) but these fail when drawn multiple times in
viewer with a non-ganesh backend (when ganesh is available). Fix this by
always using SkCanvas::baseRecorder.

[0] https://skia.googlesource.com/skia.git/+/0fef076beec39f12dc3a5bfb2a939c25627da718

Change-Id: Id904b80945222fecca54f4d3a16476e415328706
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1077076
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>