[graphite] Conditionalize call to prepareForReturnToCache
Using the Resource's shareable flag as a test for calling prepareForReturnToCache creates a race condition between when we're returning a resource to the cache and finding it in ResourceCache::findAndRefResource (in which we're setting the shareability).

This CL introduces a const class Boolean that is only true for those resources that require a prepareForReturnToCache call.

Bug: b/448374340
Change-Id: If853fe04f7ce67ebadf6275c904a12bc65d993ca
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1075936
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>