[graphite] Reduce clip data saved in DrawList
Now that uniforms are extracted ASAP in drawGeometry, the DrawList only
needs to hold state necessary for writeVertices(). The NonMSAAClip and
ClipShader aren't something that will be supported without embedding
into the PaintParamsKey, so we can reduce storage.

Specifically we are removing 64 bytes from DrawParams. Clip is at
least 112 bytes on a 64-bit machine, and now DrawParams only has to
store the 48 bytes for the 3 rectangles.

Bug: b/362716197
Change-Id: I1428974b09563aac67bbef69629984fe3c72dfdc
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1079858
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>