[skcapture] add basic capture data serialization
This CL only handles the serialization/deserialization of the header, basic metadata, and the SKPs collected during capture and is not representative of the file file format structure.

Introduces the SkCapture object that holds a lot of the serialization information to not put too much on the CaptureManager.

Bug: b/412351769

Change-Id: I85a56540077409dd1d4c33d06e0199913d206355
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1049996
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Jorge Betancourt <jmbetancourt@google.com>