[Graphite] Fix track of how feature flags affect ycbcr filters.
In Graphite we were converting VulkanYcbcrConversionInfos into
ImmutableSamplerInfo objects and then back into VulkanYcbcrConversionInfos.
During these conversions we drop the feature flags on the Ycbcr infos
since those flags technically aren't needed or used to key the real
backend Vulkan ycbcr sampler objects. However, we did use those flags for
down grading the chromeFilter and possibly the main sampler filters.

In Vulkan we first creating the ImageViews (which need the ycbcr info),
before we drop the feature flags. However, we created the Samplers after
we dropped the flags. This caused a possible mismatch in ycbcr objects
when using the image view and sampler together.

This change fixes it so we make all the decisions early when we create
VulkanYcbcrConversionInfos (forcing clients to use the ctor) and then
makes sure we track all the decisions through the conversions to and
from ImmutableSamplerInfo.

Bug: b/422502802
Bug: b/450533798
Change-Id: I4350d70e2a4a6446f24eed7d97999bd8b6926559
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1076676
Commit-Queue: Greg Daniel <egdaniel@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>