Make underlying SkData in SkMemoryStream const
There are several APIs I noticed where we take in or hold onto
a non-const sk_sp<SkData> that we do not modify and likely
should not modify. [1][2][3] Also, in SkImage::refEncodedData()
we don't want the client to modify the data we return [4].

In order to change those APIs, we need to be able to pass a const
SkData to SkMemoryStream, which should be fine given that's
a read-only view into the underlying buffer. All cases inside of
Skia that use getData() now only seem to use it in a read-only
fashion anyway, but there's a lot of changes involved with that
[5] and I want to roll them out more gradually. Thus, the introduction
of getDataView().

[1] https://skia.googlesource.com/skia/+/479696a3c5bfeb263a18c9260cb581b20ea4b556/include/private/SkEncodedInfo.h#40
[2] https://skia.googlesource.com/skia/+/479696a3c5bfeb263a18c9260cb581b20ea4b556/include/codec/SkJpegDecoder.h#34
[3] https://skia.googlesource.com/skia/+/479696a3c5bfeb263a18c9260cb581b20ea4b556/include/ports/SkTypeface_fontations.h#21
[4] https://skia.googlesource.com/skia/+/479696a3c5bfeb263a18c9260cb581b20ea4b556/include/core/SkImage.h#742
[5] https://skia-review.googlesource.com/c/skia/+/1084056

Change-Id: I138017550a05a7ac003405902adf12323f403732
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1084037
Reviewed-by: Florin Malita <fmalita@google.com>
Auto-Submit: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Florin Malita <fmalita@google.com>