[graphite] Default to SK_TRACE_GRAPHITE_PIPELINE_USE always enabled
Android is always going to want this, and we've run into roll failures
because that was the only place that had the define enabled. It also is
a feature that would be valuable for Chrome's tracing.

With both Android and Chrome enabling it, it seems worth always enabling
it. We could conditionally track this data based on whether or not
SK_TRACING is defined, but it's hopefully a low enough overhead to sum
up areas that it's negligible.

Change-Id: Ib004fd07dad3ea27df85d402990bbb290886cd68
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1085738
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>