[graphite] Drop draws with invalid PaintParamsKeys
We always dropped drawImage calls with images that couldn't be converted
to a Graphite-backed representation. But image shaders had no way of
dropping the draw when pipelines and PaintParamsKeys were determined
during DrawPass creation because everything was already accumulated.

Now that PaintParamsKeys are extracted early, we know if the SkImage
referenced by an image shader is valid or not and we can signal the
Device to drop the draw. This is extended to any other case where
KeyHelpers was adding an error code block to the key. This consistency
is important for the follow-up CL that moves drawImage's image handling
to be reliant on the PaintParamsKey building just like it were an image
shader.

Added a unit test to verify. Before this CL, the "imageShader" test
variation would fail because the draws would proceed and render the
error magenta color. With this CL, the tests pass and behave the same as
the "drawImageRect" variation.

Bug: b/238766195
Change-Id: Id2e0520966f7e0a76fd37061e970977973ff3dc1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1086457
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: Thomas Smith <thomsmit@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>