[graphite] Fix DrawWriter suballocation performance
After https://skia-review.googlesource.com/c/skia/+/1066438, the amount
of work done in each append() call was increased because each append
suballocation re-aligned the buffer's offset and checked capacity with
full validation. For path rendering, this was significant because
previously they would do a large reservation and then call append(1),
which would just do basic pointer arithmetic as long as the reservation
had elements left. The aligning and validation overhead was amortized.

On desk_micrographygirlsvg.skp nanobench, my M3 Air originally ran in
4.08ms. After the above BufferSubAllocator CL, it ran in about 5.1ms.
After this CL, it runs in 3.97ms.

This reworks the BufferSubAllocator logic so that get[Mapped]Subrange
handle aligning the offset and calculating how many strided elements
remain. Then it exposes append functions that can be called instead when
the caller knows that there will be many appends with the same stride
and alignment as before. This helped perf by about 0.75ms.

A few things that helped improve performance even past what it was: 1.
The appendMappedStride() only returns a BufferWriter and not a
   BindBufferInfo. Even if it was only using that binding when
   the pending count was 0, the memory traffic was measurable (.2ms)
2. Rearrange reserve/append() helper functions to be on DrawWriter
   and minimize the template and virtual definitions (.1ms), this also
   moves all of the vertices 4x alignment handling to the Vertices
   appender.
3. Forcing these functions to SK_ALWAYS_INLINE further helped (<.1ms but
consistent).

Bug: b/455564823
Change-Id: I8b72ebd5dceb892355a2315143fb1c680cf01456
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1085416
Reviewed-by: Thomas Smith <thomsmit@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>