Fix fuzzer targets and subsequent bitrotting
In https://review.skia.org/1057137 there was some erroneous
placement of braces that caused the fuzz targets to not be
built at all (because they were in their own definition).

This fixes that by completely removing the indentation
(in a future CL, I might move the fuzz definitions to a different
file for better organization). I've disabled building the fuzzers
by default (I think trying to build them for *every* Linux+Clang
host was a bit problematic, and made them opt-in. This is
also why I've added a pair of CI jobs (one on the CQ) to explicitly
build the fuzzers.

Also, some code had bitrotted due to the fact that we weren't
building Skia with SK_BUILD_FOR_FUZZER set (but our oss-fuzz
pipeline *was* setting this and breaking w/o us knowing). This
fixes the bitrot and moves defines around so when we enable the
fuzzers, the define is set. This also removes the
skia_use_libfuzzer_defaults setting, which wasn't really doing
anything and now allows the libfuzzer fuzzers to actually be
built w/o adding any manual cxx or linker flags.

To test this, I built Skia locally with `skia_build_fuzzers`
set to True (like the CQ/CI job does) and also patched this
into an oss-fuzz checkout and saw that it works there too.

Change-Id: I1bc6381b62402ee868e00354bd039c27ab03ff2f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1090236
Reviewed-by: Eric Boren <borenet@google.com>