[graphite] Refactor helper function for local matrix shader block adding
Moves calculating the inverse of the local matrix into the helper,
and sets the error block on the paint key builder if that fails, isntead
of switching to using the identity matrix.

Handles doing the postConcat of any extra transform (for images and
gradients), as well as then invoking the child addInner function with
the modified key context pointing to the local matrix.

For now this is used for the SkLocalMatrixShader and SkCTMShader, but
the simple-image handling will add a 3rd use of this pattern.

Change-Id: Ic539dcaceb7056e03a4d0a4b0b3579422e2c0928
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1091196
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>