[graphite] Add Caps::toString virtual
This will be used to pretty print YCbCr information in the Pipeline labels.

As an example, this will replace the substring:
  HardwareImage(3: kAwAEPcAAAAAAAAA)
with:
  HardwareImage(x247 2020+narrow cos cos nearest F rgba cf1lf0)

Unfortunately, wgpu::YCbCrVkDescriptor and skgpu::VulkanYcbcrConversionInfo have diverged (in their handling of the ChromaFilter and member variable ordering) so we can't share code between native Vulkan and Dawn Vulkan.

Bug: b/456438502
Change-Id: I3b78af7a546e30c0104cd988bb43b5f08e9886ef
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1087937
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>