Add SMPTE ST 2094-50 HDR metadata support
Add the structure skhdr::Agtm, implementing several parts (not all) of
the SMPTE ST 2094-50 spec. Add tests for the parts that have been
written.

The first candidate draft of ST 2094-50 can be viewed at:
https://github.com/SMPTE/st2094-50 The second candidate draft is not yet
published.

Move the HDR tests to a new file.

Leave this structure in a header inside src/ (rather that in
include/private/SkHdrMetadata.h), because it's likely that we will want
to expose an abstract interface in the public-ish header.

Bug: skia:456675139
Change-Id: Ib38f7e6bb4523c53f86bc3b1bb318c2d3f8cc45d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1089956
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Christopher Cameron <ccameron@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>