Add SkPathBuilder::rMoveTo(x,y)
All of the other r*To have versions which take scalers and forward to
the structured versions. Add this for rMoveTo.

Also update the r*To to take SkVector instead of SkPoint and rename the
parameters with `d` to match the documentation.

Bug: 40040287
Change-Id: I73e1cbd8c0cc44c88e1d1626dbb5a09d8ee9f8be
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1092936
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>
Auto-Submit: Ben Wagner <bungeman@google.com>
Commit-Queue: Florin Malita <fmalita@google.com>