[graphite] Use combined depth+stencil format when using MSAA
MSAA is triggered for path rendering, but for convex paths and stroked
paths, the stencil attachment is not necessary. This can lead to
scenarios where the renderpass would be D16x4 or D24S8x4 depending on
whether or not more complex paths were included.

This causes a 3x recompile for single-sample-requiring pipelines
instead of just 2x if we always use D+Sx4. Similarly, it causes a 2x
recompile for the convex and stroked paths instead of just 1x if
we adopt D+Sx4.

It can theoretically cause an increase in attachment memory if
content were to only need the D16x4 attachment (D24S8 is twice as big).
However, in practice we see chromium requiring tiles that are D16x4
and other tiles that are D24S8x4. If we only requested D24S8x4, we
create one MSAA'ed attachment that gets reused for both tiles, in
which case we have an improvement of 50% over needing both a
D16x4 and D24S8x4 attachment.

Bug: b/458732319
Change-Id: Ia00343ac8a1f02fa66b74668426e095506201a33
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1095857
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>