[graphite] Move PathRendererStrategy to RendererProvider, remove kDefault
The end goal is Caps can have a requested strategy from its private
context options in test builds, but RendererProvider stores the chosen
strategy used for all path rendering in the current context.

This moves the enum from ContextOptionsPriv.h to RendererProvider.h and
removes the kDefault option. Instead the requested strategy is wrapped
in an optional so that it can override the default selection process
when set by viewer.

This also fixes some of the radio button initialization logic in Viewer
where it was adding the "Software" button when in a graphite config, and
adding the Ganesh buttons when in a software config because there was
still a ganesh context for rendering the UI.

Bug: b/452351267
Change-Id: Ia251ca987bc12a9adc37de778b0d6895b19e604a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1096777
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Thomas Smith <thomsmit@google.com>