[Graphite] Conditionally initialize ClipAtlasManager
Use the presence of the ClipAtlasManager to determine whether or not
excess clip elements are sent to the atlas or to the depth-only
rasterization.

This removes the chosen path strategy from ClipStack's behavior,
although AtlasProvider uses the same decision on strategy to decide
whether or not to allocate the ClipAtlasManager.

Now we avoid allocating the manager when it's never to be used, and
in the future, if we ever want to use it for other strategies, it's
easy to adjust.

Bug: b/452351267
Change-Id: Ic654dce678b5dceaa55999b31a22ab9af8284387
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1100358
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>