[graphite] Simplify code for SW rasterizing paths/clips to pixmaps.
There was unnecessary use of SkAutoPixmapStorage for rendering to atlases. The end product for cases that did need to allocate is SkBitmap, so just use that.

Extra offsets and bounds were being passed around to make and use
RasterMaskHelper instead of making snug pixmaps.

Change-Id: Ie65157509ed536b0159b6d673c555bcd022b6823
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1084036
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Brian Salomon <briansalomon@gmail.com>
Reviewed-by: Thomas Smith <thomsmit@google.com>