[pdf] Only paint non-degenenerate paths
In some cases degenerate paths or verbs may be discarded (like when
filling) and in other cases they should be emitted or preserved (like
when clipping or stroking). Make these cases clear in the code. Allow
users to know when path data was actually emitted so that empty paths
can be handled appropriately.

Bug: 460715107
Change-Id: I2c193b04c6a9c208a49f843643c8f0fd363a273e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1103476
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>