AGTM: Add metadata parsing support
Add parsing for the current version of the metadata structure, based on
the current draft (which is subject to change).

The SMPTE ST 2094-50 draft is divided into two sections:
* syntax specifies the bitstream
* semantics specifies the mapping of syntax elements to metadata items

The skhdr::Agtm structure stores the metadata items. Add a separate
private AgtmSyntax structure to store the syntax. The parsing and
serialization functions are done in two steps.
* going from/to a SkStream to/from an AgtmSyntax
* going from/to an AgtmSyntax to/from an skhdr::Agtm

Add tests of serialize<->deserialize, but not using hardcoded data just
yet (that will wait until the format is certain not to change any more).

Fuzzing tests need to be added.

Bug: skia:456675139
Change-Id: I594c553fc00fa077bd5d6e35efd83182a02d3023
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1100336
Reviewed-by: Christopher Cameron <ccameron@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>