Fix quoting of windows paths with spaces
Apparently, quoting the paths when passed into cmake defines like
CMAKE_C_COMPILER is not helpful, as it throws of CMake's parsing
(these were likely left over from earlier debugging of something).

We *do* need to quote the paths passed in for include paths, which
this now does. I also cleaned up some mixes of \ and / that can
happen (not that it broke anything, but it certainly looked odd
during debugging).

I tested this locally with GN args
    clang_win="C:\\cipd\\clang win with spaces"
    win_sdk="C:\\cipd\\win toolchain with spaces\\win_sdk"
    win_vc="C:\\cipd\\win toolchain with spaces\\VC"

And it worked! CMake saw this correctly
-- C++ compiler.........: C:/cipd/clang win with spaces/bin/clang-cl.exe

Hint for next time, the logs mentioned a CMakeLists.txt and that
was implied to be the one in third_party/dawn/externals (even
though the full path wasn't on the stacktrace)

Bug: b/456384551
Change-Id: Ifc7e6b7c2fe66d2d0d4226bd13873e6221236e62
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1105856
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Eric Boren <borenet@google.com>