Make SkImage and SkImageGenerator return const SkData from refEncoded
This behavior was implied (a shared pointer to SkData should always
be treated as read-only) and is now made explicit. This also updates
several places in our decoders where we promise not to modify the
SkData that is passed in (since we will likely be caching it) and
ask that the client not modify it either.

This is partially guarded by SK_DISABLE_LEGACY_NONCONST_ENCODED_IMAGE_DATA
so we can update Chromium and potentially other places.

Change-Id: Id174dff4afa148745ad32dd7d5cd848931b77f9d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1084056
Reviewed-by: Florin Malita <fmalita@google.com>