SkEncodedInfo::ICCProfile: Rename to ColorProfile
Rename SkEncodedInfo::ICCProfile to SkCodecs::ColorProfile, and move it
to a private header.

Add a wide range of Make functions for all of the ways that it is
initialized by the various codecs.

The next steps are to create a SkCodecs::ColorTransform structure (that
will hold the details of the at-decode-time transforms), and then to
move all use of skcms and its types (excepting skcms_TransferFunction
and skcms_Matrix3x3) to be behind these interfaces.

Bug: skia:456675139
Change-Id: If1643d1d4ff4f2f1fecae25655bacfd3c1b05ddf
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1112036
Reviewed-by: Christopher Cameron <ccameron@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>